+++ packages/client/src/components/var/VarPrompt.tsx
@@ -62,7 +62,7 @@ export const VarPrompt = ({
 }: IVarBaseInputProps<string>): JSX.Element => {
   const [currentValue, setCurrentValue, currentError] = useVarUIValue({
     path,
-    fallbackValue: value,
+    fallbackValue: value || '',
     onChange,
     error,
     errorPath,